Add ending soon alert
