+++ app/client/main.scss
@@ -21,6 +21,8 @@ i.material-icons {
 @import "{}/imports/ui/pages/settings/settings.scss";
 
 /* Components */
+@import "{}/imports/ui/components/alerts/alert-ending-soon/alert-ending-soon.scss";
+
 @import "{}/imports/ui/components/footer/footer.scss";
 @import "{}/imports/ui/components/nav/nav.scss";
