+++ app/imports/ui/components/alerts/alert-ending-soon/alert-ending-soon.scss
@@ -1,0 +1,4 @@
+.alert-ending-soon a {
+  color: currentColor;
+  font-weight: bold;
+}
