+++ app/imports/ui/components/queue-actions/queue-actions.scss
@@ -1,4 +1,5 @@
 .queue-actions {
+  border-top: 1px solid $gray-lighter;
   border-bottom: 1px solid $gray-lighter;
 
   .container {
