+++ app/imports/ui/components/queue-alert-restricted-session/queue-alert-restricted-session.scss
@@ -1,5 +1,5 @@
 .queue-alert-restricted-session {
-  border-bottom: 1px solid $gray-lighter;
+  border-top: 1px solid $gray-lighter;
 }
 
 .device-cards {
