+++ app/imports/ui/components/queue-header/queue-header.scss
@@ -1,7 +1,3 @@
-.queue-header {
-  border-bottom: 1px solid $gray-lighter;
-}
-
 .queue-header-background {
   position: relative;
