+++ app/imports/ui/pages/queue/queue.html
@@ -2,6 +2,7 @@
   <div class="queue">
     {{#if Template.subscriptionsReady}}
       {{> QueueHeader queue=queue taView=taView}}
+      {{> AlertEndingSoon queue=queue taView=taView}}
 
       {{#if queue.isRestricted}}
         {{> QueueAlertRestrictedSession queue=queue taView=taView}}
