+++ app/imports/ui/pages/queue/queue.js
@@ -6,6 +6,7 @@ import { Roles } from 'meteor/alanning:roles';
 import { Queues } from '/imports/api/queues/queues.js';
 
 import '/imports/ui/components/queue-header/queue-header.js';
+import '/imports/ui/components/alerts/alert-ending-soon/alert-ending-soon.js';
 import '/imports/ui/components/queue-actions/queue-actions.js';
 import '/imports/ui/components/queue-alert-restricted-session/queue-alert-restricted-session.js';
 import '/imports/ui/components/queue-table/queue-table.js';
